User management

+++ routes/web.php
@@ -11,6 +11,8 @@
 |
 */
 
+use App\Role;
+
 Route::get('/', function () {
     return view('welcome');
 });
@@ -20,12 +22,12 @@ Auth::routes();
 Route::get('/', 'CategoryController@index')->name('categories');
 
 Route::get('/categories', 'CategoryController@index')->name('categories');
-Route::post('/categories', 'CategoryController@insert')->name('insertCategory');
-Route::get('/categories/add', 'CategoryController@add')->name('addCategory');
-Route::get('/categories/edit/{id}', 'CategoryController@edit')->name('editCategory');
+Route::post('/categories', 'CategoryController@insert')->name('insertCategory')->middleware('checkRole:' . Role::ROLES['ADMIN']);
+Route::get('/categories/add', 'CategoryController@add')->name('addCategory')->middleware('checkRole:' . Role::ROLES['ADMIN']);
+Route::get('/categories/edit/{id}', 'CategoryController@edit')->name('editCategory')->middleware('checkRole:' . Role::ROLES['ADMIN']);
 Route::get('/categories/{id}', 'CategoryController@show')->name('showCategory');
-Route::delete('/categories/{id}', 'CategoryController@delete')->name('deleteCategory');
-Route::put('/categories/{id}', 'CategoryController@update')->name('updateCategory');
+Route::delete('/categories/{id}', 'CategoryController@delete')->name('deleteCategory')->middleware('checkRole:' . Role::ROLES['ADMIN']);
+Route::put('/categories/{id}', 'CategoryController@update')->name('updateCategory')->middleware('checkRole:' . Role::ROLES['ADMIN']);
 
 
 Route::post('/subcategories', 'SubCategoryController@insert')->name('insertSubCategory');
@@ -33,3 +35,7 @@ Route::get('/subcategories/edit/{id}', 'SubCategoryController@edit')->name('edit
 Route::get('/subcategories/{id}', 'SubCategoryController@show')->name('showSubCategory');
 Route::delete('/subcategories/{id}', 'SubCategoryController@delete')->name('deleteSubCategory');
 Route::put('/subcategories/{id}', 'SubCategoryController@update')->name('updateSubCategory');
+
+Route::get('/users', 'UserController@index')->name('users')->middleware('checkRole:' . Role::ROLES['ADMIN']);
+Route::put('/users/{id}', 'UserController@update')->name('updateUser')->middleware('checkRole:' . Role::ROLES['ADMIN']);
+
